requirements.txt
python-dotenv
lightning
numpy
pandas
pytorch-forecasting
scikit-learn
torch
torchvision
torchsummary